Begin styling header,set padding for sections, add background colors
This commit adds background colors to the webpage's sections according
to The Odin Project's Landing Page outline. A padding of 200px was
applied to the left and right of the major sections of the webpage as
shown in the outline. Removed text decoration from the links in the
header and removed all list styling for the header. I also used flexbox
to stick the footer to the bottom of the webpage.