Remove padding from some elements, finish quote section
This commit fixes an issue with the webpage where adding styling
within the quote section with flexbox would cause elements to go into
other sections of the webpage. This was fixed by removing padding
rules to some elements that were added in the previous commit. The
styling for the quote section of the webpage was also completed
according to The Odin Project's landing page outline.