Add css styling to sign up section of webpage
This commit creates a sign up section of the webpage. A blue
background was added to the box in the sign up section and the font size
was increased to 18px to make the text easier to read. Flexbox was
used in order to position the sign up button to the right side of the
section.